modifiche

+++ src/main/java/backend_esame4/Dispositivi_Aziendali/model/Dispositivo.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class Dispositivo {
     @Id
     @GeneratedValue
-    private long id;
+    private int id;
     @Enumerated(EnumType.STRING)
     private StatoDispositivo statoDispositivo;
 
@@ -21,4 +21,3 @@ public class Dispositivo {
     private Dipendente dipendente;
 
 }
-
